feat: 平台菜单/统计/CORS 后端 + 登录页/仪表盘前端（此前未提交 WIP）
后端：PlatformMenu 一整套(entity/mapper/service/controller/dto/vo/xml)、
StatisticsController、CorsConfig，及 CaptchaConfig/MenuController/
WebSocketAutoConfiguration/TenantCacheService/init.sql/platform.sql/
两个 application.yml/pom.xml 等小幅调整。

前端：登录页 authentication.vue 重构(+294)、仪表盘/概览、
新组件 particle-network/type-writer/directives、preferences/locales/pnpm-lock。

注：本批为会话前已存在的未提交改动，未逐行审查，完整性/可编译性未验证。

.gitignore: 忽略 sh.exe.stackdump / *.stackdump。

+++ services/platform-service/src/main/java/com/saas/cloud/platform/service/impl/OrderServiceImpl.java
@@ -141,7 +141,7 @@ public class OrderServiceImpl extends ServiceImpl<OrderMapper, Order> implements
 
         tenant.setPackageId(order.getPackageId());
         tenant.setPaidExpireTime(expireTime);
-        tenant.setStatus(1);
+        tenant.setStatus((byte) 1);
         tenantService.updateById(tenant);
 
         order.setExpireTime(expireTime);
